requirements.txt
flask
flask-cors
openai
faiss-cpu
gunicorn
pandas
